First pass at "working tests"
