+++ server/mediapublic/development.ini
@@ -12,7 +12,6 @@ pyramid.debug_notfound = false
 pyramid.debug_routematch = false
 pyramid.default_locale_name = en
 pyramid.includes =
-    pyramid_debugtoolbar
     pyramid_tm
 
 sqlalchemy.url = sqlite:///%(here)s/mediapublic.sqlite
