+++ server/mediapublic/mediapublic/__init__.py
@@ -14,7 +14,6 @@ def main(global_config, **settings):
     DBSession.configure(bind=engine)
     Base.metadata.bind = engine
     config = Configurator(settings=settings)
-    config.include('pyramid_chameleon')
     config.include('cornice')
     config.scan('mediapublic.views')
