+++ server/mediapublic/test-requirements.txt
@@ -1,2 +1,3 @@
 pep8
 gabbi
+nose
