+++ server/mediapublic/tox.ini
@@ -1,0 +1,28 @@
+[tox]
+minversion = 2.0
+envlist = py27,py33,py34,pep8
+
+[testenv]
+commands = nosetests
+setenv = VIRTUAL_ENV={envdir}
+         PYTHONHASHSEED=0
+deps = -r{toxinidir}/requirements.txt
+       -r{toxinidir}/test-requirements.txt
+
+[testenv:pep8]
+commands = pep8
+
+[testenv:venv]
+commands = {posargs}
+
+[testenv:cover]
+commands = python setup.py test --coverage --testr-args='{posargs}'
+
+[testenv:gabbi]
+commands = python -m unittest -v mediapublic/tests/api.py
+deps = gabbi
+
+[pep8]
+#ignore = H305,H307,H402,H904
+show-source = True
+exclude =  .venv,.git,.tox,dist,docs,*egg,build,*setup.py
